requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Cors==4.0.0
Flask-Redis==0.4.0
Flask-Session==0.8.0
authlib==1.3.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.1
Werkzeug==3.0.1
psycopg2-binary==2.9.9